# Conflicts:
#	src/main/resources/.env.properties

+++ src/main/java/com/app/salty/board/controller/ArticleController.java
@@ -1,18 +1,29 @@
 package com.app.salty.board.controller;
 
 import com.app.salty.board.dto.article.*;
+import com.app.salty.board.entity.Article;
+import com.app.salty.board.entity.ArticleHeader;
 import com.app.salty.board.service.ArticleServiceImpl;
+import com.app.salty.user.entity.CustomUserDetails;
 import com.app.salty.user.entity.Users;
+import com.app.salty.user.service.UserService;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.io.File;
 import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Paths;
 import java.util.List;
+import java.util.UUID;
 
+@Slf4j
 @RestController
 @RequestMapping("/api")
 public class ArticleController {
@@ -21,6 +32,75 @@ public class ArticleController {
     ArticleController(ArticleServiceImpl articleService) {
         this.articleService = articleService;
     }
+// ====================================================================================================================
+
+    private final String uploadDir = Paths.get("C:", "tui-editor", "upload").toString();
+
+    /**
+     * 에디터 이미지 업로드
+     * @param image 파일 객체
+     * @return 업로드된 파일명
+     */
+    @PostMapping("/image-upload")
+    public String uploadEditorImage(@RequestParam final MultipartFile image) {
+        if (image.isEmpty()) {
+            return "";
+        }
+
+        String orgFilename = image.getOriginalFilename();                                         // 원본 파일명
+        String uuid = UUID.randomUUID().toString().replaceAll("-", "");           // 32자리 랜덤 문자열
+        assert orgFilename != null;
+        String extension = orgFilename.substring(orgFilename.lastIndexOf(".") + 1);  // 확장자
+        String saveFilename = uuid + "." + extension;                                             // 디스크에 저장할 파일명
+        String fileFullPath = Paths.get(uploadDir, saveFilename).toString();                      // 디스크에 저장할 파일의 전체 경로
+
+        // uploadDir에 해당되는 디렉터리가 없으면, uploadDir에 포함되는 전체 디렉터리 생성
+        File dir = new File(uploadDir);
+        if (!dir.exists()) {
+            dir.mkdirs();
+        }
+
+        try {
+            // 파일 저장 (write to disk)
+            File uploadFile = new File(fileFullPath);
+            image.transferTo(uploadFile);
+            return saveFilename;
+
+        } catch (IOException e) {
+            // 예외 처리는 따로 해주는 게 좋습니다.
+            throw new RuntimeException(e);
+        }
+    }
+
+    /**
+     * 디스크에 업로드된 파일을 byte[]로 반환
+     * @param filename 디스크에 업로드된 파일명
+     * @return image byte array
+     */
+    @GetMapping(value = "/image-print", produces = { MediaType.IMAGE_GIF_VALUE, MediaType.IMAGE_JPEG_VALUE, MediaType.IMAGE_PNG_VALUE })
+    public byte[] printEditorImage(@RequestParam final String filename) {
+        // 업로드된 파일의 전체 경로
+        String fileFullPath = Paths.get(uploadDir, filename).toString();
+
+        // 파일이 없는 경우 예외 throw
+        File uploadedFile = new File(fileFullPath);
+        if (!uploadedFile.exists()) {
+            throw new RuntimeException();
+        }
+
+        try {
+            // 이미지 파일을 byte[]로 변환 후 반환
+            return Files.readAllBytes(uploadedFile.toPath());
+
+        } catch (IOException e) {
+            // 예외 처리는 따로 해주는 게 좋습니다.
+            throw new RuntimeException(e);
+        }
+    }
+
+    //====================================================================================================================
+
+
 
     // 게시물 전체 조회
     @GetMapping("/article")
@@ -37,33 +117,30 @@ public class ArticleController {
     }
 
     // 게시물 저장
-    @PostMapping(value = "/article", consumes ={MediaType.APPLICATION_JSON_VALUE
-            , MediaType.MULTIPART_FORM_DATA_VALUE})
-    public ResponseEntity<SaveArticleResponseDto> saveArticle(@RequestPart SaveArticleRequestDto requestDto
-            , @RequestPart("files") MultipartFile[] files) throws IOException {
-
-        // 임의의 유저 생성 - test
-        Users tempUser = new Users();
-        tempUser.setId(1L);
-        // ==============================
-
-        requestDto.setUser(tempUser);
-        return ResponseEntity.status(HttpStatus.CREATED).body(articleService.saveArticle(requestDto, files));
+    @PostMapping(value = "/article")
+    public ResponseEntity<SaveArticleResponseDto> saveArticle(@RequestBody SaveArticleRequestDto requestDto
+            ,@AuthenticationPrincipal CustomUserDetails user){
+        requestDto.setUserId(user.getId());
+        return ResponseEntity.status(HttpStatus.CREATED).body(articleService.saveArticle(requestDto));
     }
 
     // 게시물(articleId) 수정
-    @PutMapping("/article/{articleId}")
-    public ResponseEntity<UpdateArticleResponseDto> updateArticle(@PathVariable Long articleId
-            , @RequestBody UpdateArticleRequestDto requestDto
-            , @AuthenticationPrincipal Users user) {
+    @PutMapping(value ="/article/{articleId}")
+    public ResponseEntity<UpdateArticleResponseDto> updateArticle(@RequestBody UpdateArticleRequestDto requestDto
+            , @PathVariable Long articleId) {
 
-        return ResponseEntity.ok(articleService.updateArticle(requestDto, articleId));
+        return ResponseEntity.ok(articleService.updateArticle(requestDto));
     }
 
     // 게시물(articleId) 삭제
     @DeleteMapping("/article/{articleId}")
     public ResponseEntity<Void> deleteArticle(@PathVariable Long articleId
-            ,  @AuthenticationPrincipal Users user) {
+            ,  @AuthenticationPrincipal CustomUserDetails user) {
+        Long writerID = articleService.getArticleById(articleId).getWriterId();
+        if(!writerID.equals(user.getId())) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
         articleService.deleteArticle(articleId);
         return ResponseEntity.ok().build();
     }
@@ -76,8 +153,9 @@ public class ArticleController {
 
     // 게시물(articleId)과 해당 게시물의 댓글 함께 조회
     @GetMapping("/article/comment/{articleId}")
-    public ResponseEntity<GetArticleWithCommentResponseDto> getArticleWithComment(@PathVariable Long articleId) {
-        GetArticleWithCommentResponseDto dtoList = articleService.getArticleWithCommentByArticleId(articleId);
+    public ResponseEntity<GetArticleWithCommentResponseDto> getArticleWithComment(@PathVariable Long articleId
+            ,@AuthenticationPrincipal CustomUserDetails user) {
+        GetArticleWithCommentResponseDto dtoList = articleService.getArticleWithCommentByArticleId(articleId, user);
         return ResponseEntity.ok(dtoList);
     }
